+++ TFG/JAVA/Model/src/main/java/org/iesfm/model/services/ItemService.java
@@ -1,8 +1,12 @@
 
 package org.iesfm.model.services;
 
+import org.iesfm.model.pojos.Category;
 import org.iesfm.model.pojos.Item;
+import org.iesfm.model.pojos.Supplier;
+import org.iesfm.model.repositories.CategoryRepository;
 import org.iesfm.model.repositories.ItemRepository;
+import org.iesfm.model.repositories.SupplierRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,15 +17,26 @@ public class ItemService {
 
     @Autowired
     private ItemRepository itemRepository;
+    @Autowired
+    private CategoryRepository categoryRepository;
+    @Autowired
+    private SupplierRepository supplierRepository;
 
     public ItemService(ItemRepository itemRepository) {
         this.itemRepository = itemRepository;
     }
 
-    public Item postItem(Item newItem){
-        if (itemRepository.existsById(newItem.getId())){
+    public Item postItem(Item newItem) {
+        if (itemRepository.existsById(newItem.getId())) {
             return null;
-        }else{
+        } else if (categoryRepository.findByName(newItem.getCategory()) == null){
+            Category category = new Category(0, newItem.getCategory());
+            categoryRepository.save(category);
+            itemRepository.save(newItem);
+            return newItem;
+        } else if (supplierRepository.findByName(newItem.getSupplier()) == null) {
+            return null;
+        } else {
             itemRepository.save(newItem);
             return newItem;
         }
@@ -31,26 +46,33 @@ public class ItemService {
         return itemRepository.findAll();
     }
 
-    public Item getItem(Integer id) {
+    public Item getItem(int id) {
         if (!itemRepository.existsById(id)) {
             return null;
-        }else {
-            return itemRepository.getById(id);
+        } else {
+            return itemRepository.findById(id).get();
         }
     }
 
-    public Item putItem(Integer id, Item itemUpdated) {
+    public Item putItem(int id, Item itemUpdated) {
+        if (!itemRepository.existsById(id)) {
+            return null;
+        } else {
             Item item = getItem(id);
             itemRepository.delete(item);
             itemUpdated.setId(id);
             itemRepository.save(itemUpdated);
             return itemUpdated;
+        }
     }
 
-    public Item deleteItem(Integer id) {
-        Item itemDeleted = getItem(id);
-        itemRepository.delete(itemDeleted);
-        return itemDeleted;
+    public Item deleteItem(int id) {
+        if (!itemRepository.existsById(id)) {
+            return null;
+        } else {
+            itemRepository.delete(itemRepository.findById(id).get());
+            return getItem(id);
+        }
     }
 
 }
